requirements.txt
mysql_connector_repackaged==0.3.1
psycopg2==2.9.6
python-dotenv==1.0.0
